# swagger.json
{"openapi":"3.0.0","paths":{"/v1/events":{"get":{"operationId":"EventController_getEvents","summary":"Return a list of avaliable events.","parameters":[{"name":"Content-Type","in":"header","description":"application/json","schema":{"type":"string"}}],"responses":{"200":{"description":"List of avaliable events","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventListDto"}}}},"500":{"description":"Error not handled","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"internalServerError"},"status":{"type":"number","example":500},"type":{"example":"/v1/events"},"detail":{"type":"string","example":"unexpected error"},"errors":{"type":"array","example":[]}}}}}}},"tags":["Event"]},"post":{"operationId":"EventController_createEvent","summary":"Create an event.","parameters":[{"name":"Authorization","required":true,"in":"header","description":"Bearer accessToken","schema":{"type":"string","example":"Bearer accessToken"}},{"name":"Content-Type","in":"header","description":"application/json","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateEventDto"}}}},"responses":{"201":{"description":"Event successfully created","content":{"application/json":{"schema":{"type":"object","example":{}}}}},"400":{"description":"Some inputs are missed or wrong","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"badRequestError"},"status":{"type":"number","example":400},"type":{"example":"/v1/events"},"detail":{"type":"string","example":"some inputs are missing"},"errors":{"type":"array","example":["location.name should not be empty"]}}}}}},"401":{"description":"Error for invalid token","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"unauthorizedRequestError"},"status":{"type":"number","example":401},"type":{"example":"/v1/events"},"detail":{"type":"string","example":"not authorized"},"errors":{"type":"array","example":["token invalid"]}}}}}},"403":{"description":"Error for not have enough privileges","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"unauthorizedRequestError"},"status":{"type":"number","example":403},"type":{"example":"/v1/events"},"detail":{"type":"string","example":"not authorized"},"errors":{"type":"array","example":["not enough privileges"]}}}}}},"500":{"description":"Error not handled","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"internalServerError"},"status":{"type":"number","example":500},"type":{"example":"/v1/events"},"detail":{"type":"string","example":"unexpected error"},"errors":{"type":"array","example":[]}}}}}}},"tags":["Event"],"security":[{"bearer":[]}]},"patch":{"operationId":"EventController_updateEvent","summary":"Update an event.","parameters":[{"name":"Authorization","required":true,"in":"header","description":"Bearer accessToken","schema":{"type":"string","example":"Bearer accessToken"}},{"name":"Content-Type","in":"header","description":"application/json","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateEventDto"}}}},"responses":{"200":{"description":"Event successfully updated","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventDto"}}}},"400":{"description":"Error for not existing event","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"badRequestError"},"status":{"type":"number","example":400},"type":{"example":"/v1/events"},"detail":{"example":"eventId"},"errors":{"example":["eventId invalid"]}}}}}},"401":{"description":"Error for invalid token","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"unauthorizedRequestError"},"status":{"type":"number","example":401},"type":{"example":"/v1/events"},"detail":{"type":"string","example":"not authorized"},"errors":{"type":"array","example":["token invalid"]}}}}}},"403":{"description":"Error for not have enough privileges","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"unauthorizedRequestError"},"status":{"type":"number","example":403},"type":{"example":"/v1/events"},"detail":{"type":"string","example":"not authorized"},"errors":{"type":"array","example":["not enough privileges"]}}}}}},"500":{"description":"Error not handled","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"internalServerError"},"status":{"type":"number","example":500},"type":{"example":"/v1/events"},"detail":{"type":"string","example":"unexpected error"},"errors":{"type":"array","example":[]}}}}}}},"tags":["Event"],"security":[{"bearer":[]}]},"delete":{"operationId":"EventController_deleteEvent","summary":"Delete event and the participant list related","parameters":[{"name":"Authorization","required":true,"in":"header","description":"Bearer accessToken","schema":{"type":"string","example":"Bearer accessToken"}},{"name":"Content-Type","in":"header","description":"application/json","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventIdDto"}}}},"responses":{"204":{"description":"Event data has been succesfully delete"},"400":{"description":"Some inputs are missed or wrong","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"badRequestError"},"status":{"type":"number","example":400},"type":{"example":"/v1/events"},"detail":{"type":"string","example":"some inputs are missing"},"errors":{"type":"array","example":["eventId should not be empty"]}}}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"unauthorizedRequestError"},"status":{"type":"number","example":401},"type":{"type":"string","example":"/v1/events"},"detail":{"type":"string","example":"not authorized"},"errors":{"type":"array","example":["token invalid"]}}}}}},"403":{"description":"Invalid credentials","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"unauthorizedRequestError"},"status":{"type":"number","example":403},"type":{"example":"/v1/events"},"detail":{"type":"string","example":"not authorized"},"errors":{"type":"array","example":["not enough privileges"]}}}}}},"500":{"description":"Error not handled","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"internalServerError"},"status":{"type":"number","example":500},"type":{"type":"string","example":"/v1/events"},"detail":{"type":"string","example":"unexpected error"},"errors":{"type":"array","example":[]}}}}}}},"tags":["Event"],"security":[{"bearer":[]}]}},"/v1/account/events":{"get":{"operationId":"AccountController_getMyEvents","summary":"Return a list of events the user joined and have not been celebrated","parameters":[{"name":"Authorization","required":true,"in":"header","description":"Bearer accessToken","schema":{"type":"string","example":"Bearer accessToken"}}],"responses":{"200":{"description":"List of avaliable events user joined","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventListDto"}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"unauthorizedRequestError"},"status":{"type":"number","example":401},"type":{"type":"string","example":"/v1/account/events"},"detail":{"type":"string","example":"not authorized"},"errors":{"type":"array","example":["token invalid"]}}}}}},"500":{"description":"Error not handled","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"internalServerError"},"status":{"type":"number","example":500},"type":{"type":"string","example":"/v1/account/events"},"detail":{"type":"string","example":"unexpected error"},"errors":{"type":"array","example":[]}}}}}}},"tags":["Account"]},"post":{"operationId":"AccountController_joinEvent","summary":"Allow authenticated users to join as participants in an event","parameters":[{"name":"Authorization","required":true,"in":"header","description":"Bearer accessToken","schema":{"type":"string","example":"Bearer accessToken"}},{"name":"Content-Type","in":"header","description":"application/json","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventIdDto"}}}},"responses":{"204":{"description":"User has successfully joined in the event"},"400":{"description":"Some inputs are missed or wrong","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"badRequestError"},"status":{"type":"number","example":400},"type":{"type":"string","example":"/v1/account/events"},"detail":{"type":"string","example":"some inputs are missing"},"errors":{"type":"array","example":["eventId should not be empty"]}}}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"unauthorizedRequestError"},"status":{"type":"number","example":401},"type":{"type":"string","example":"/v1/account/events"},"detail":{"type":"string","example":"not authorized"},"errors":{"type":"array","example":["token invalid"]}}}}}},"406":{"description":"Can not join celebrated events","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"notAcceptable"},"status":{"type":"number","example":406},"detail":{"type":"string","example":"not acceptable"},"type":{"type":"string","example":"/v1/account/events"},"errors":{"type":"array","example":["cannot join past events"]}}}}}},"500":{"description":"Error not handled","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"internalServerError"},"status":{"type":"number","example":500},"type":{"type":"string","example":"/v1/account/events"},"detail":{"type":"string","example":"unexpected error"},"errors":{"type":"array","example":[]}}}}}}},"tags":["Account"],"security":[{"bearer":[]}]},"delete":{"operationId":"AccountController_disJoinEvent","summary":"Allow authenticated users to disjoin as participants from an event","parameters":[{"name":"Authorization","required":true,"in":"header","description":"Bearer accessToken","schema":{"type":"string","example":"Bearer accessToken"}},{"name":"Content-Type","in":"header","description":"application/json","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventIdDto"}}}},"responses":{"204":{"description":"User has successfully joined in the event"},"400":{"description":"Some inputs are missed or wrong","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"badRequestError"},"status":{"type":"number","example":400},"type":{"type":"string","example":"/v1/account/events"},"detail":{"type":"string","example":"some inputs are missing"},"errors":{"type":"array","example":["eventId should not be empty"]}}}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"unauthorizedRequestError"},"status":{"type":"number","example":401},"type":{"type":"string","example":"/v1/account/events"},"detail":{"type":"string","example":"not authorized"},"errors":{"type":"array","example":["token invalid"]}}}}}},"406":{"description":"Can not disjoin celebrated events","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"notAcceptable"},"status":{"type":"number","example":406},"detail":{"type":"string","example":"not acceptable"},"type":{"type":"string","example":"/v1/account/events"},"errors":{"type":"array","example":["cannot disjoin past events"]}}}}}},"500":{"description":"Error not handled","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"internalServerError"},"status":{"type":"number","example":500},"type":{"type":"string","example":"/v1/account/events"},"detail":{"type":"string","example":"unexpected error"},"errors":{"type":"array","example":[]}}}}}}},"tags":["Account"],"security":[{"bearer":[]}]}},"/v1/account/login":{"post":{"operationId":"AccountController_login","summary":"Login an user.","parameters":[{"name":"Content-Type","in":"header","description":"application/json","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountLoginDto"}}}},"responses":{"200":{"description":"User has successfully logged in","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountLoggedDto"}}}},"400":{"description":"Some inputs are missed or wrong","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"badRequestError"},"status":{"type":"number","example":400},"type":{"type":"string","example":"/v1/account/login"},"detail":{"type":"string","example":"some inputs are missing"},"errors":{"type":"array","example":["email should not be empty"]}}}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"unauthorizedRequestError"},"status":{"type":"number","example":401},"type":{"type":"string","example":"/v1/account/login"},"detail":{"type":"string","example":"inputs are invalid"},"errors":{"type":"array","example":["email password combination is not correct"]}}}}}},"500":{"description":"Error not handled","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"internalServerError"},"status":{"type":"number","example":500},"type":{"type":"string","example":"/v1/account/login"},"detail":{"type":"string","example":"unexpected error"},"errors":{"type":"array","example":[]}}}}}}},"tags":["Account"]}},"/v1/account/register":{"post":{"operationId":"AccountController_register","summary":"Register a new user","parameters":[{"name":"Content-Type","in":"header","description":"application/json","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountRegisterDto"}}}},"responses":{"201":{"description":"User has successfully created and verification email has been sent","content":{"application/json":{"schema":{"type":"object","example":{}}}}},"400":{"description":"Some inputs are missed or wrong","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"badRequestError"},"status":{"type":"number","example":400},"type":{"type":"string","example":"/v1/account/register"},"detail":{"type":"string","example":"some inputs are missing"},"errors":{"type":"array","example":["email should not be empty"]}}}}}},"500":{"description":"Error not handled","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"internalServerError"},"status":{"type":"number","example":500},"type":{"type":"string","example":"/v1/account/login"},"detail":{"type":"string","example":"unexpected error"},"errors":{"type":"array","example":[]}}}}}}},"tags":["Account"]}},"/v1/account/refresh":{"post":{"operationId":"AccountController_refreshToken","summary":"Refresh token.","parameters":[{"name":"Content-Type","in":"header","description":"application/json","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"200":{"description":"A new pair of accesToken and refreshToken is succesfully provided","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewPairOfTokensDto"}}}},"400":{"description":"Some inputs are missed or wrong","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"badRequestError"},"status":{"type":"number","example":400},"type":{"type":"string","example":"/v1/account/refresh"},"detail":{"type":"string","example":"some inputs are missing"},"errors":{"type":"array","example":["refreshToken should not be empty"]}}}}}},"401":{"description":"Invalid credentials","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"unauthorizedRequestError"},"status":{"type":"number","example":401},"type":{"type":"string","example":"/v1/account/refresh"},"detail":{"type":"string","example":"unauthorized"},"errors":{"type":"array","example":["unauthorized"]}}}}}},"500":{"description":"Error not handled","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"internalServerError"},"status":{"type":"number","example":500},"type":{"type":"string","example":"/v1/account/refresh"},"detail":{"type":"string","example":"unexpected error"},"errors":{"type":"array","example":[]}}}}}}},"tags":["Account"]}},"/v1/account/activate":{"put":{"operationId":"AccountController_activate","summary":"Update account status to ACTIVE for user the verificationCode is related","parameters":[{"name":"Content-Type","in":"header","description":"application/json","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActivateAccountDto"}}}},"responses":{"200":{"description":"User status succesfully updated to ACTIVE","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountLoggedDto"}}}},"400":{"description":"Some inputs are missed or wrong","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"badRequestError"},"status":{"type":"number","example":400},"type":{"type":"string","example":"/v1/account/activate"},"detail":{"type":"string","example":"some inputs are missing"},"errors":{"type":"array","example":["verificationCode should not be empty"]}}}}}},"500":{"description":"Error not handled","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"internalServerError"},"status":{"type":"number","example":500},"type":{"type":"string","example":"/v1/account/activate"},"detail":{"type":"string","example":"unexpected error"},"errors":{"type":"array","example":[]}}}}}}},"tags":["Account"]}},"/v1/account/password/forgot-init":{"post":{"operationId":"AccountController_forgotPassword","summary":"Allow users to send an email with a code to request the change of a forgotten password.","parameters":[{"name":"Content-Type","in":"header","description":"application/json","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailDto"}}}},"responses":{"204":{"description":"The email for forgotten password has been sent successfully"},"400":{"description":"Some inputs are missed or wrong","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"badRequestError"},"status":{"type":"number","example":400},"type":{"type":"string","example":"/v1/account/password/forgot-init"},"detail":{"type":"string","example":"some inputs are missing"},"errors":{"type":"array","example":["email should not be empty"]}}}}}},"500":{"description":"Error not handled","content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","example":"internalServerError"},"status":{"type":"number","example":500},"type":{"type":"string","example":"/v1/account/password/forgot-init"},"detail":{"type":"string","example":"unexpected error"},"errors":{"type":"array","example":[]}}}}}}},"tags":["Account"]}}},"info":{"title":"Esmorga API","description":"[Swagger in JSON format](http://localhost:3000/swagger-json)\n\n Swagger for Esmorga API.","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3000"}],"components":{"schemas":{"LocationDto":{"type":"object","properties":{"lat":{"type":"number","example":43.35525182148881,"maximum":90,"minimum":-90,"description":"GPS Latitude"},"long":{"type":"number","example":-8.41937931298951,"maximum":180,"minimum":-180,"description":"GPS Longitude"},"name":{"type":"string","example":"A Coruña","minLength":1,"maxLength":100}},"required":["name"]},"EventDto":{"type":"object","properties":{"eventId":{"type":"string","example":"6656e23640e1fdb4ceb23cc9","minLength":24,"maxLength":24},"eventName":{"type":"string","example":"MobgenFest","minLength":3,"maxLength":100},"eventDate":{"format":"date-time","type":"string","example":"2024-03-08T10:05:30.915Z"},"description":{"type":"string","example":"Hello World","minLength":4,"maxLength":5000},"eventType":{"type":"string","example":"Party","enum":["Party","Sport","Food","Charity","Games"]},"imageUrl":{"type":"string","example":"img.url","maxLength":500},"location":{"$ref":"#/components/schemas/LocationDto"},"tags":{"example":["Dance","Music"],"type":"array","items":{"type":"string","maxLength":25,"minLength":3}}},"required":["eventId","eventName","eventDate","description","eventType","location"]},"EventListDto":{"type":"object","properties":{"totalEvents":{"type":"number","example":1},"events":{"type":"array","items":{"$ref":"#/components/schemas/EventDto"}}},"required":["totalEvents","events"]},"CreateEventLocationDto":{"type":"object","properties":{"name":{"type":"string","example":"A Coruña","minLength":1,"maxLength":100},"lat":{"type":"number","example":43.35525182148881,"maximum":90,"minimum":-90,"description":"GPS Latitude"},"long":{"type":"number","example":-8.41937931298951,"maximum":180,"minimum":-180,"description":"GPS Longitude"}},"required":["name"]},"CreateEventDto":{"type":"object","properties":{"eventName":{"type":"string","example":"End of the World Party","minLength":3,"maxLength":100},"eventDate":{"type":"string","example":"2025-03-08T10:05:30.915Z","format":"date-time"},"description":{"type":"string","example":"Join us for an unforgettable celebration as we dance into the apocalypse.","minLength":4,"maxLength":5000},"eventType":{"type":"string","example":"Party","enum":["Party","Sport","Food","Charity","Games"]},"imageUrl":{"type":"string","example":"image.url","maxLength":500},"location":{"$ref":"#/components/schemas/CreateEventLocationDto"},"tags":{"example":["Dance","Music"],"type":"array","items":{"type":"string","maxLength":25,"minLength":3}}},"required":["eventName","eventDate","description","eventType","location"]},"UpdateEventLocationDto":{"type":"object","properties":{"name":{"type":"string","example":"A Coruña","minLength":1,"maxLength":100},"lat":{"type":"number","example":43.35525182148881,"maximum":90,"minimum":-90,"description":"GPS Latitude"},"long":{"type":"number","example":-8.41937931298951,"maximum":180,"minimum":-180,"description":"GPS Longitude"}}},"UpdateEventDto":{"type":"object","properties":{"eventId":{"type":"string","example":"6656e23640e1fdb4ceb23cc9","minLength":24,"maxLength":24},"eventName":{"type":"string","example":"End of the World Party","minLength":3,"maxLength":100},"eventDate":{"type":"string","example":"2025-03-08T10:05:30.915Z","format":"date-time"},"description":{"type":"string","example":"Join us for an unforgettable celebration as we dance into the apocalypse.","minLength":4,"maxLength":5000},"eventType":{"type":"string","example":"Party","enum":["Party","Sport","Food","Charity","Games"]},"imageUrl":{"type":"string","example":"image.url","maxLength":500},"location":{"$ref":"#/components/schemas/UpdateEventLocationDto"},"tags":{"example":["Dance","Music"],"type":"array","items":{"type":"string","maxLength":25,"minLength":3}}},"required":["eventId"]},"EventIdDto":{"type":"object","properties":{"eventId":{"type":"string","example":"6670301dfcb094d1f8f3ea8d"}},"required":["eventId"]},"AccountLoginDto":{"type":"object","properties":{"email":{"type":"string","example":"esmorga.test.01@yopmail.com","maxLength":100},"password":{"type":"string","example":"Password01","minLength":8,"maxLength":50}},"required":["email","password"]},"AccountLoggedDto":{"type":"object","properties":{"accessToken":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1dWlkIjoiNjY1ZjAxOWMxNzMzMWViZWU1NTBiMmZkIiwiaWF0IjoxNzE4MDE2NDMzLCJleHAiOjE3MTgwMTcwMzN9.migGA-6g8dCJ_4i6-AGQHN4girESXQmn5MtM9U7NhJg"},"refreshToken":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1dWlkIjoiNjY1ZjAxOWMxNzMzMWViZWU1NTBiMmZkIiwiaWF0IjoxNzE4MDE2NDMzfQ.CXgzuTrEI38H02-f_Jv3LEa_W5K6pAGXwVIDf6902S8"},"ttl":{"type":"number","example":"600"},"profile":{"type":"object","properties":{"name":{"type":"string","example":"John","minLength":3,"maxLength":100},"lastName":{"type":"string","example":"Donnel-Vic","minLength":3,"maxLength":100},"email":{"type":"string","example":"eventslogin01@yopmail.com","maxLength":100}}}},"required":["accessToken","refreshToken","ttl","profile"]},"AccountRegisterDto":{"type":"object","properties":{"name":{"type":"string","example":"John","minLength":3,"maxLength":100},"lastName":{"type":"string","example":"Donnel-Vic","minLength":3,"maxLength":100},"email":{"type":"string","example":"eventslogin01@yopmail.com","maxLength":100},"password":{"type":"string","example":"SuperSecret1!","minLength":8,"maxLength":50}},"required":["name","lastName","email","password"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1dWlkIjoiNjY1ZjAxOWMxNzMzMWViZWU1NTBiMmZkIiwiaWF0IjoxNzE4MDE2NDMzfQ.CXgzuTrEI38H02-f_Jv3LEa_W5K6pAGXwVIDf6902S"}},"required":["refreshToken"]},"NewPairOfTokensDto":{"type":"object","properties":{"accessToken":{"type":"string","example":"eyJhbGciOiJIUzI1NIkpXVCJ9.eyJ1dWlkIjoiNjY1ZjAxOWMxiMmZkIiwiaWF0IjoxNzE4MDE2NDMzLCJlcwMzN9.migGA-6g8dCJ_4i6-AGQHN4girESXQmn5MtM9U7NhJg"},"refreshToken":{"type":"string","example":"eyJhbGciOiJIUzcCI6IkpXVCJ9.eyJ1dWlkIjoiNjY1ZjAxOWMxNzMzMWViZWUjoxNzE4MDE2NDMzfQ.CXgzu8H02-f_Jv3LEa_W5K6pAG2S8"},"ttl":{"type":"number","example":"600"}},"required":["accessToken","refreshToken","ttl"]},"ActivateAccountDto":{"type":"object","properties":{"verificationCode":{"type":"string","example":"123456","minLength":6,"maxLength":6}},"required":["verificationCode"]},"EmailDto":{"type":"object","properties":{"email":{"type":"string","example":"eventslogin01@yopmail.com","maxLength":100}},"required":["email"]}}}}